# openapi.json
{"openapi":"3.1.0","info":{"title":"Student API Services","description":"Student API Documentation","termsOfService":"company url !!","version":"1.0"},"servers":[{"url":"http://localhost:8080/api","description":"Generated server url"}],"tags":[{"name":"Student","description":"APIs for managing students (Create, Read, Update, Delete)"}],"paths":{"/students/{id}":{"get":{"tags":["Student"],"summary":"Get student by ID","description":"Fetches a student using their unique ID","operationId":"getStudentById","parameters":[{"name":"id","in":"path","description":"Unique ID of the student","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"404":{"description":"Student not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"Student found successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StudentDto"}}}}}},"put":{"tags":["Student"],"summary":"Update student (Full)","description":"Completely updates an existing student","operationId":"updateStudent","parameters":[{"name":"id","in":"path","description":"ID of the student to update","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddStudentRequestDto"}}},"required":true},"responses":{"404":{"description":"Student not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"409":{"description":"Email already exists","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"Student updated successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StudentDto"}}}}}},"delete":{"tags":["Student"],"summary":"Delete student","description":"Deletes a student using their ID","operationId":"deleteAStudent","parameters":[{"name":"id","in":"path","description":"ID of the student to delete","required":true,"schema":{"type":"integer","format":"int64"},"example":101}],"responses":{"404":{"description":"Student not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"204":{"description":"Student deleted successfully"}}},"patch":{"tags":["Student"],"summary":"Update student (Partial)","description":"Updates selected fields of an existing student","operationId":"updatePartialStudent","parameters":[{"name":"id","in":"path","description":"ID of the student to update","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"requestBody":{"description":"Fields to update (key-value pairs)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddStudentRequestDto"}}},"required":true},"responses":{"404":{"description":"Student not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"409":{"description":"Email already exists","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"400":{"description":"Invalid field provided","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"Student partially updated successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StudentDto"}}}}}}},"/students":{"get":{"tags":["Student"],"summary":"Get all students","description":"Fetches the list of all registered students","operationId":"getAllStudents","responses":{"200":{"description":"Blank list return []","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StudentDto"}}}}}}},"post":{"tags":["Student"],"summary":"Create a new student","description":"Creates and saves a new student in the system","operationId":"createNewStudent","requestBody":{"description":"Student creation request payload","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddStudentRequestDto"}}},"required":true},"responses":{"409":{"description":"Email already exists","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"201":{"description":"Student created successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StudentDto"}}}},"400":{"description":"Validation failed","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}}}}}},"components":{"schemas":{"AddStudentRequestDto":{"type":"object","description":"Request payload for creating a new student","properties":{"name":{"type":"string","description":"Full name of the student (3 to 30 characters)","example":"Rishabh","maxLength":30,"minLength":3},"email":{"type":"string","format":"email","description":"Valid email address of the student","example":"rishabh@gmail.com"}},"required":["email","name"]},"ApiError":{"type":"object","description":"Standard structure for API error responses","properties":{"status":{"type":"integer","format":"int32","description":"HTTP status code of the error","example":404},"message":{"type":"string","description":"Human-readable description of the error","example":"Student not found with id : 10"},"timestamp":{"type":"string","format":"date-time","description":"Timestamp when the error occurred","example":"2026-02-03T16:45:12"}}},"StudentDto":{"type":"object","description":"Represents student information returned by the API","properties":{"id":{"type":"integer","format":"int64","description":"Unique identifier of the student","example":1,"readOnly":true},"name":{"type":"string","description":"Name of the student","example":"Rishabh"},"email":{"type":"string","format":"email","description":"Email address of the student","example":"rishabh@gmail.com"}},"required":["email","name"]}}}}